Fix call check_vendor_specific_psustatus error (#2016)
* fix call check_vendor_specific_psustatus error

* move check_vendor_specific_psustatus to test_platform_info.py
